backend/requirements.txt
alembic==1.16.1
blinker==1.9.0
click==8.2.1
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.20.1
packaging==25.0
PyJWT==2.10.1
PyMySQL==1.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
Werkzeug==2.3.7
